+++ src/redux/action-creators.js
@@ -1273,6 +1273,22 @@ export function enableBansInGroup(groupName) {
   };
 }
 
+export function getPostsByIds(postIds) {
+  return {
+    apiRequest: Api.getPostsByIds,
+    type: ActionTypes.GET_POSTS_BY_IDS,
+    payload: { postIds },
+  };
+}
+
+export function getCommentsByIds(commentIds) {
+  return {
+    apiRequest: Api.getCommentsByIds,
+    type: ActionTypes.GET_COMMENTS_BY_IDS,
+    payload: { commentIds },
+  };
+}
+
 export function setCurrentRoute(payload) {
   return {
     type: ActionTypes.SET_CURRENT_ROUTE,