+++ src/redux/reducers.js
@@ -709,7 +709,8 @@ export function comments(state = {}, action) {
     case response(ActionTypes.SHOW_MORE_COMMENTS):
     case response(ActionTypes.COMPLETE_POST_COMMENTS):
     case response(ActionTypes.GET_SINGLE_POST):
-    case response(ActionTypes.SHOW_MORE_LIKES_ASYNC): {
+    case response(ActionTypes.SHOW_MORE_LIKES_ASYNC):
+    case response(ActionTypes.GET_COMMENTS_BY_IDS): {
       return updateCommentData(state, action);
     }
     case response(ActionTypes.GET_COMMENT_BY_NUMBER): {